requirements.txt
argparse == 1.4.0
pillow == 10.2.0
